requirements.txt
altgraph==0.17.2
autopep8==1.6.0
bidict==0.21.4
bleach==4.1.0
cachelib==0.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cmake==3.22.1
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
dnspython==2.1.0
docutils==0.18.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SocketIO==4.3.1
flaskwebgui==0.3.5
fonttools==4.28.5
future==0.18.2
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
imutils==0.5.4
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.4.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.21.5
opencv-python==4.5.4.60
packaging==21.3
pefile==2022.5.30
Pillow==8.4.0
pkginfo==1.8.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
pyparsing==3.0.6
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
pywin32-ctypes==0.2.0
readme-renderer==32.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
spotipy==2.19.0
toml==0.10.2
tqdm==4.62.3
twine==3.7.1
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
whichcraft==0.6.1
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
